Only set upgrade flag and reboot after the upgrade command has been run successfully

+++ tpx-clj/upgrader.sh
@@ -5,15 +5,17 @@ output_dir=$1
 echo "stopping services"
 systemctl stop sp-tpx sp-bridgeprogram
 
-# set upgrading flag that'll be picked up by TPX on next boot
-echo "Setting upgrading_flag"
-echo "upgrading" > "$output_dir"/upgrading_flag
-
 apt-get update
 
 # perform the upgrade
 echo "upgrading"
 apt-get install -y --only-upgrade teleporter-fw
+status=$?
+if $(exit $status); then
+    # set upgrading flag that'll be picked up by TPX on next boot
+    echo "Setting upgrading_flag"
+    echo "upgrading" > "$output_dir"/upgrading_flag
 
-echo "rebooting"
-reboot
+    echo "rebooting"
+    reboot
+fi
